requirements.txt
Django>=5.1,<5.2
wagtail>=6.3,<6.4
python-decouple